+++ src/main/java/cn/uptra/schoolwork/modules/book/service/ReviewService.java
@@ -8,19 +8,19 @@ import java.util.List;
 public interface ReviewService extends IService<Review> {
 
     // 从 mapper 抽象到 service：根据 bid 返回书本的平均评分
-    Double getAverageRatingByBid(Integer bid);
+    Double getAverageRatingByBid(Long bid);
 
     // 根据 bid 返回书本的所有评论数量
-    Integer getReviewCountByBid(Integer bid);
+    Integer getReviewCountByBid(Long bid);
 
     // 根据 bid 和 uid 查询用户是否已评论该书
-    Integer hasUserReviewedBook(Integer bid, Long uid);
+    Integer hasUserReviewedBook(Long bid, Long uid);
 
     // 根据 bid 和 uid 查询用户的评论内容
-    Review getUserReviewForBook(Integer bid, Long uid);
+    Review getUserReviewForBook(Long bid, Long uid);
 
     // 根据 bid 返回书本的所有评论，按时间排序
-    List<Review> getAllReviewsByBid(Integer bid);
+    List<Review> getAllReviewsByBid(Long bid);
 
     // 根据 uid 返回用户的所有评论
     List<Review> getAllReviewsByUid(Long uid);
